backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1